Use a dedicated badge for the assertible/deployments repo
The current badge on this repo is using the Assertible service's
status. Sometimes we have (unrelated) test failures on that service and
it shouldn't affect the assertible/deployments repo badge.

I created a service specifically for assertible/deployments and updated
the badge in the README to use that service's badge which should always
be passing.